chore(swiper): add swiper-bundle.min.css file
The commit adds the `swiper-bundle.min.css` file to the `static/lib/swiper` directory. This file contains the CSS styles for the Swiper library version 11.0.3. The styles define the appearance and behavior of the Swiper slider component. The file is released under the MIT License and was last updated on October 26, 2023.